requirements.txt
Django==2.2
requests==2.24.0
django-storages==1.11.1
dropbox==11.10.0
pytz~=2021.3
djangorestframework==3.11.0
python-decouple==3.3
gunicorn==20.1.0
whitenoise==5.2.0
